# web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "identity": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Identity",
            "type": "string"
          }
        },
        "required": [
          "identity"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "SettingsActionRequest": {
        "properties": {
          "confirmation": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Confirmation",
            "type": "string"
          },
          "section": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "SettingsActionRequest",
        "type": "object"
      },
      "SettingsAuditRequest": {
        "properties": {
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "operation"
        ],
        "title": "SettingsAuditRequest",
        "type": "object"
      },
      "SettingsAuthorizationRequest": {
        "properties": {
          "operation": {
            "default": "settings.save",
            "title": "Operation",
            "type": "string"
          },
          "permission": {
            "default": "settings.edit",
            "title": "Permission",
            "type": "string"
          }
        },
        "title": "SettingsAuthorizationRequest",
        "type": "object"
      },
...
      "SettingsWriteRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "value"
        ],
        "title": "SettingsWriteRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EOAT Atlas API",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/development/login": {
      "post": {
        "operationId": "development_login_api_v1_auth_development_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Development Login"
      }
    },
...
    "/api/v1/settings/actions/{action}": {
      "post": {
        "operationId": "settings_action_api_v1_settings_actions__action__post",
        "parameters": [
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings Action"
      }
    },
    "/api/v1/settings/audit": {
      "post": {
        "operationId": "audit_settings_action_api_v1_settings_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Settings Action"
      }
    },
    "/api/v1/settings/authorization/check": {
      "post": {
        "operationId": "authorize_settings_api_v1_settings_authorization_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsAuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Authorize Settings Api V1 Settings Authorization Check Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorize Settings"
      }
    },
...
    "/api/v1/settings/{setting_key}": {
      "put": {
        "operationId": "write_setting_api_v1_settings__setting_key__put",
        "parameters": [
          {
            "in": "path",
            "name": "setting_key",
            "required": true,
            "schema": {
              "title": "Setting Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Setting"
      }
    },
...
  }
}